pyproject.toml
[project]
name = "opassit"
version = "0.2.0"
description = "企业运营与自动化工作平台 - Django + MySQL"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "OPASSIT Team" }]
keywords = ["django", "enterprise", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "Django>=5.0,<6.0",
    "pymysql>=1.1.0",
    "cryptography>=42.0.0",
    "python-dotenv>=1.0.0",
    "django-cors-headers>=4.3.0",
    "gunicorn>=21.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
